+++ lib/generators/templates/helpers/amp/meta_tag_helper.rb
@@ -1,9 +1,0 @@
-module ApplicationHelper
-
-    def meta_tag(options: {})
-        options = options.symbolize_keys
-
-        render "meta", options: options
-    end
-
-end